Assignment 2: Refactor to MVC architecture
