+++ config/app.js
@@ -1,5 +1,5 @@
 /*
-File name: app.js
+File name: config/app.js
 Student name: Cong Lanh Hoang
 Student ID: 301210743
 Date: September 30, 2022
@@ -13,22 +13,22 @@ var cookieParser = require("cookie-parser"); // Parse Cookie header and populate
 var logger = require("morgan"); // HTTP logger
 
 // import routers
-var indexRouter = require("./routes/index");
-var usersRouter = require("./routes/users");
+var indexRouter = require("../routes/index.router");
+var usersRouter = require("../routes/user.router");
 
 // instantiate new express object
 var app = express();
 
 // view engine setup
-app.set("views", path.join(__dirname, "views")); // set the location of views folder
+app.set("views", path.join(__dirname, "../views")); // set the location of views folder
 app.set("view engine", "ejs"); // Set ejs as engine to render view template
 
 app.use(logger("dev")); // use logger to log HTTP request
 app.use(express.json()); // express render json
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, "public"))); // set static folder to public folder
-app.use(express.static(path.join(__dirname, "node_modules"))); // second static folder, contain bootstrap and fontawesome
+app.use(express.static(path.join(__dirname, "../public"))); // set static folder to public folder
+app.use(express.static(path.join(__dirname, "../node_modules"))); // second static folder, contain bootstrap and fontawesome
 
 app.use("/", indexRouter); // defined paths after root path
 app.use("/users", usersRouter); // defined paths after users root path
