+++ controllers/user.controller.js
@@ -1,0 +1,10 @@
+/*
+File name: controllers/index.controller.js
+Student name: Cong Lanh Hoang
+Student ID: 301210743
+Date: October 14, 2022
+*/
+
+exports.userList = function (req, res, next) {
+  res.send("respond with a resource");
+};
