--- routes/index.js
@@ -1,36 +1,0 @@
-/*
-File name: routes/index.js
-Student name: Cong Lanh Hoang
-Student ID: 301210743
-Date: Oct 5, 2022
-*/
-
-var express = require("express");
-var router = express.Router();
-
-/* GET home page. */
-router.get("/", function (req, res, next) {
-  res.render("index", { title: "Home" });
-});
-
-/* GET About Me page. */
-router.get("/about-me", function (req, res, next) {
-  res.render("index", { title: "About Me" });
-});
-
-/* GET Projects page. */
-router.get("/projects", function (req, res, next) {
-  res.render("index", { title: "Projects" });
-});
-
-/* GET Services page. */
-router.get("/services", function (req, res, next) {
-  res.render("index", { title: "Services" });
-});
-
-/* GET Contact page. */
-router.get("/contact", function (req, res, next) {
-  res.render("index", { title: "Contact" });
-});
-
-module.exports = router;
