+++ server.js
@@ -4,7 +4,7 @@
  * Module dependencies.
  */
 
-var app = require("../app");
+var app = require("./config/app");
 var debug = require("debug")("comp229.018.assignment1:server");
 var http = require("http");
